Fix CWE-681: correct numeric type conversions in 8 files
- ActiveDirectoryDateUtil: Math.toIntExact() for long→int guard
- PasswordGenerator: remove spurious (float) intermediate casts
- Drupal7Encoder: integer ceiling division replaces float Math.ceil
- AbstractEncoder: explicit (byte)(int) two-step cast
- PBEWithMD5AndDESEncoder: bounds check before byte cast of salt length
- Sha256Crypt/Sha512Crypt: nextInt() replaces (int)(nextFloat()*N)
- Util.randomSleep: Math.round(double) returns long, no cast needed

+++ identity4j-active-directory-jndi/src/main/java/com/identity4j/connector/jndi/activedirectory/ActiveDirectoryDateUtil.java
@@ -58,7 +58,8 @@ public final class ActiveDirectoryDateUtil {
      */
     public static int adTimeToJavaDays(long timeStamp) {
         // CWE-570 fix: compare the full long, not a truncated (int) cast
-        return timeStamp == Long.MIN_VALUE ? 0 : (int) (timeStamp / -86400L / 10000000L);
+        // CWE-681 fix: use Math.toIntExact to detect overflow instead of silent truncation
+        return timeStamp == Long.MIN_VALUE ? 0 : Math.toIntExact(timeStamp / -86400L / 10000000L);
     }
 
 	public static long javaDateToADTime(Date date) {
